+++ examples/hello_rant.rs
@@ -2,8 +2,8 @@ use rant::Rant;
 use std::error::Error;
 
 fn main() -> Result<(), Box<dyn Error>> {
-  // Create a default Rant context and load the standard library
-  let mut rant = Rant::new();
+  // Create a Rant context and load the standard library
+  let mut rant = Rant::with_random_seed();
 
   // Compile a simple program
   let program = rant.compile_quiet(r#"