+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 4.0.0-alpha.5
 
+* Added `Rant::with_random_seed()` constructor
 * Added support for multi-part accessors
 * Upgraded `quickscope` to 0.1.2